+++ Program.cs
@@ -32,7 +32,22 @@ static class Program
 //  + Review ARCHITECTURE.md and validate that it matches current implementation
 //  + Improve ARCHITECTURE.md: Make sure that infrastructure layer is on the diagram
 //  + Выбирать в варианты "похожие" слова - нужно определять "расстояние" между словами
-//  ** Возможность просить напечатать ответ (требуется нормализация списка)
+//  ++ Возможность просить напечатать ответ (требуется нормализация списка)
+//   + IMPROVEMENTS:
+//     + Show correcly typed characters in hint (if hint is turned on)
+//     + Parametrize tests (for now only for TypingQuizPresenter)
+//     + Validate single responsibility principle for TypingQuizPresenter
+//     + Switch to Typing difficulty when reveal is set
+//     + After correct answer close window on "enter"
+//     + If answer has brackets, do not take them into account
+//     + In LetterHintTracker, if gate closed and not locked, should be a chance to reveal a bonus letter for the first option
+//     + Algorithm in SequenceAligner.FindMatches should be properly described in javadoc (or whatever they call it)
+//     + "Bonus" reveal should show random symbol, not one by one
+//  * Review namespace and directory structure
+//  * Refine words list and make sure that it is compliant
+//    * Exclude words with non-latin letters or with more than 3 words in answer (including article and words in brackets)
+//    * All multioption answers should be put correctly
+//    * Questions should not contain commas or brackets
 
 
 //todo Написать скрипт, чтобы смержить words и appdata, привести финальный лист в нормальный вид, убрать дубликаты если есть