+++ ViewModels/OptionsViewModel.cs
@@ -1,6 +1,5 @@
 using System;
 using VocabularyTrainer.Models;
-using VocabularyTrainer.Services;
 
 namespace VocabularyTrainer.ViewModels;
 
@@ -9,8 +8,7 @@ namespace VocabularyTrainer.ViewModels;
 /// </summary>
 public class OptionsViewModel : ViewModelBase
 {
-    private readonly SettingsService _settingsService;
-    private readonly Action _onSettingsApplied;
+    private readonly Action<int, int, int, QuizDirection, QuizDifficulty, bool> _onSave;
     private readonly Action _onClosed;
 
     private int _quizIntervalSeconds;
@@ -21,6 +19,8 @@ public class OptionsViewModel : ViewModelBase
     private bool _isRandomMode;
     private bool _isEasyMode;
     private bool _isHardMode;
+    private bool _isTypingMode;
+    private bool _isTypingRevealLetters;
 
     /// <summary>
     /// Gets or sets the quiz interval in seconds.
@@ -82,7 +82,11 @@ public class OptionsViewModel : ViewModelBase
     public bool IsEasyMode
     {
         get => _isEasyMode;
-        set => SetProperty(ref _isEasyMode, value);
+        set
+        {
+            SetProperty(ref _isEasyMode, value);
+            if (value) IsTypingRevealLetters = false;
+        }
     }
 
     /// <summary>
@@ -91,7 +95,38 @@ public class OptionsViewModel : ViewModelBase
     public bool IsHardMode
     {
         get => _isHardMode;
-        set => SetProperty(ref _isHardMode, value);
+        set
+        {
+            SetProperty(ref _isHardMode, value);
+            if (value) IsTypingRevealLetters = false;
+        }
+    }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether Typing difficulty is selected.
+    /// </summary>
+    public bool IsTypingMode
+    {
+        get => _isTypingMode;
+        set => SetProperty(ref _isTypingMode, value);
+    }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether letters should be revealed on wrong attempts in Typing mode.
+    /// </summary>
+    public bool IsTypingRevealLetters
+    {
+        get => _isTypingRevealLetters;
+        set
+        {
+            SetProperty(ref _isTypingRevealLetters, value);
+            if (value && !_isTypingMode)
+            {
+                IsEasyMode = false;
+                IsHardMode = false;
+                IsTypingMode = true;
+            }
+        }
     }
 
     /// <summary>
@@ -107,24 +142,22 @@ public class OptionsViewModel : ViewModelBase
     /// <summary>
     /// Initializes a new instance of the <see cref="OptionsViewModel"/> class.
     /// </summary>
-    /// <param name="settingsService">The settings service.</param>
-    /// <param name="onSettingsApplied">Callback when settings are saved.</param>
+    /// <param name="initialSettings">The current settings to populate the form.</param>
+    /// <param name="onSave">Callback invoked with the new settings values when saving.</param>
     /// <param name="onClosed">Callback when the window is closed.</param>
-    public OptionsViewModel(SettingsService settingsService, Action onSettingsApplied, Action onClosed)
+    public OptionsViewModel(AppSettings initialSettings, Action<int, int, int, QuizDirection, QuizDifficulty, bool> onSave, Action onClosed)
     {
-        _settingsService = settingsService ?? throw new ArgumentNullException(nameof(settingsService));
-        _onSettingsApplied = onSettingsApplied ?? throw new ArgumentNullException(nameof(onSettingsApplied));
+        _onSave = onSave ?? throw new ArgumentNullException(nameof(onSave));
         _onClosed = onClosed ?? throw new ArgumentNullException(nameof(onClosed));
 
-        LoadFromSettings();
+        LoadFromSettings(initialSettings ?? throw new ArgumentNullException(nameof(initialSettings)));
 
         SaveCommand = new RelayCommand(SaveAndClose);
         CancelCommand = new RelayCommand(_onClosed);
     }
 
-    private void LoadFromSettings()
+    private void LoadFromSettings(AppSettings settings)
     {
-        var settings = _settingsService.GetSettings();
         QuizIntervalSeconds = settings.QuizIntervalSeconds;
         AutoCloseSeconds = settings.QuizConfiguration.AutoCloseAfterCorrectSeconds;
         OptionCount = settings.QuizConfiguration.OptionCount;
@@ -142,8 +175,13 @@ public class OptionsViewModel : ViewModelBase
                 break;
         }
 
-        IsEasyMode = settings.QuizConfiguration.Difficulty == QuizDifficulty.Easy;
-        IsHardMode = settings.QuizConfiguration.Difficulty == QuizDifficulty.Hard;
+        switch (settings.QuizConfiguration.Difficulty)
+        {
+            case QuizDifficulty.Easy:   IsEasyMode = true; break;
+            case QuizDifficulty.Hard:   IsHardMode = true; break;
+            case QuizDifficulty.Typing: IsTypingMode = true; break;
+        }
+        IsTypingRevealLetters = settings.QuizConfiguration.TypingRevealLetters;
     }
 
     private QuizDirection GetSelectedDirection()
@@ -156,19 +194,13 @@ public class OptionsViewModel : ViewModelBase
     }
 
     private QuizDifficulty GetSelectedDifficulty() =>
-        IsHardMode ? QuizDifficulty.Hard : QuizDifficulty.Easy;
+        IsHardMode   ? QuizDifficulty.Hard   :
+        IsTypingMode ? QuizDifficulty.Typing :
+        QuizDifficulty.Easy;
 
     private void SaveAndClose()
     {
-        _settingsService.UpdateSettings(
-            QuizIntervalSeconds,
-            AutoCloseSeconds,
-            OptionCount,
-            GetSelectedDirection(),
-            GetSelectedDifficulty()
-        );
-
-        _onSettingsApplied();
+        _onSave(QuizIntervalSeconds, AutoCloseSeconds, OptionCount, GetSelectedDirection(), GetSelectedDifficulty(), IsTypingRevealLetters);
         _onClosed();
     }
 }